[MODIFY] id int to long

+++ data/src/main/java/com/findapple/findapple/data/features/post/dto/response/FindListResponse.kt
@@ -9,7 +9,7 @@ data class FindListResponse(
     val category: String,
     val detail: String,
     val findAt: String,
-    val findId: Int,
+    val findId: Long,
     val findImages: List<String>,
     val findUser: String,
     val kakaoId: Long,
